Supporting chacha20-ietf.
And using a new package to provide `chacha20` cipher, with higher performance.

old:
```
$ go test -test.bench=Chacha20Encrypt -test.benchtime=10s
BenchmarkChacha20Encrypt-28      1000000             16875 ns/op
PASS
ok      github.com/lixin9311/shadowsocks-go/shadowsocks 17.079s
```

new:
```
$ go test -test.bench=Chacha20Encrypt -test.benchtime=10s
BenchmarkChacha20Encrypt-28      5000000              2689 ns/op
PASS
ok      github.com/lixin9311/shadowsocks-go/shadowsocks 16.201s
```

about `chacha20-ietf`
#144 :
```
$ go test -test.bench=Chacha20IETFEncrypt -test.benchtime=10s
BenchmarkChacha20IETFEncrypt-28          5000000              2413 ns/op
PASS
ok      github.com/lixin9311/shadowsocks-go/shadowsocks 14.545s
```

new:
```
$ go test -test.bench=Chacha20IETFEncrypt -test.benchtime=10s
BenchmarkChacha20IETFEncrypt-28          5000000              2688 ns/op
PASS
ok      github.com/lixin9311/shadowsocks-go/shadowsocks 16.201s
```

not too much different,
to keep consistency, I personally prefer using this package.